#1 Add Forthcoming and Selected Works in Progress sections
Created a "Forthcoming" section and moved the JEP Soda tax paper into it.  Replaced "Solicited for Journal of Economic Perspectives" with "Journal of Economic Perspectives."

Added "Generalized Capital Taxation with Sufficient Statistics" and "What Is the Optimal Implicit Lottery Tax?" papers to end of the "Selected Works in Progress" section in the order specified in the issue #1.

Implemented the same changes on the papers.html and research.html files.